Fix  cUSDCv3 transfer uint256 max issue

+++ contracts/core/TermMaxMarket.sol
@@ -199,7 +199,7 @@ contract TermMaxMarket is ITermMaxMarket, ReentrancyGuard, Ownable, Pausable {
      * @inheritdoc ITermMaxMarket
      */
     function provideLiquidity(
-        uint256 underlyingAmt
+        uint128 underlyingAmt
     )
         external
         isOpen
