+++ test/Evacuate.t.sol
@@ -75,7 +75,7 @@ contract EvacuateTest is Test {
         uint amount = 10000e8;
         res.underlying.mint(deployer, amount);
         res.underlying.approve(address(res.market), amount);
-        res.market.provideLiquidity(amount);
+        res.market.provideLiquidity(uint128(amount));
 
         vm.stopPrank();
 
@@ -83,7 +83,7 @@ contract EvacuateTest is Test {
 
         vm.startPrank(sender);
         res.underlying.approve(address(res.market), amount);
-        res.market.provideLiquidity(amount);
+        res.market.provideLiquidity(uint128(amount));
         vm.stopPrank();
     }
