+++ test/Redeem.t.sol
@@ -77,7 +77,7 @@ contract RedeemTest is Test {
         uint amount = 10000e8;
         res.underlying.mint(deployer, amount);
         res.underlying.approve(address(res.market), amount);
-        res.market.provideLiquidity(amount);
+        res.market.provideLiquidity(uint128(amount));
 
         vm.stopPrank();
 
@@ -85,7 +85,7 @@ contract RedeemTest is Test {
 
         vm.startPrank(sender);
         res.underlying.approve(address(res.market), amount);
-        res.market.provideLiquidity(amount);
+        res.market.provideLiquidity(uint128(amount));
         vm.stopPrank();
     }
